ADD auto-connect to connect page
When the user presses the connect button, the app will first try to
auto connect to the device, and if that doesn’t work, it’ll return an
array of available devices that will be automatically displayed to the
screen.

This update relies on a listPorts method that I added to the
openbci-sdk.