reorganized experiments into local and remote
the experiments tab are now organized into a local and remote.  local
experiments will come with the app, and can be loaded in from the local
computer.  the remote tab will allow the user to connect to a remote
database to retrieve experiments.  the remote database is not set up
yet.